feat: add manual MTG summary trigger
