+++ README.md
@@ -203,6 +203,8 @@ Discord には以下 3 セクションを日本語で投稿します。
 当日・翌日に有効な予約が 0 件の場合は各セクションで `なし` を明示します。  
 Supabase の env 不足やクエリ失敗時も、`DBアクセス状況` に `Supabase 未照会` または `Supabase 取得失敗` を明記します。
 
+管理画面 `/appointment/admin` には、同じサマリーを即時送信する手動実行ボタンもあります。
+
 ## 検証コマンド
 
 ```bash
